chore: move abstract classes to base folder
